moved cg solver to solvers.py
